+++ doc/conf.py
@@ -26,7 +26,7 @@ author = 'J. A. Corbal'
 # The short X.Y version
 version = ''
 # The full version, including alpha/beta/rc tags
-release = '1.3.1b2'
+release = '1.3.1b3'
 
 
 # -- General configuration ---------------------------------------------------