+++ pynfact/__main__.py
@@ -17,4 +17,3 @@ from pynfact.main import main
 
 if __name__ == '__main__':
     main()
-