+++ pynfact/builder.py
@@ -9,14 +9,14 @@ Build the content.
     Include processing of reStructuredText files along with Markdown.
     Those files are detected by extension and parsed accordingly.  So,
     no extension test in this module, only in the parser.
-
 """
 from datetime import datetime
 from feedgen.feed import FeedGenerator
 from jinja2 import Environment, FileSystemLoader, Markup
 from math import ceil
+from pynfact.dataman import update_list_in_dict
 from pynfact.fileman import link_to
-from pynfact.fileman import has_extension_md_rst as valid_ext
+from pynfact.fileman import has_extension_md_rst
 from pynfact.meta import Meta
 from pynfact.parser import Parser
 from pynfact.struri import slugify, strip_html_tags
@@ -60,6 +60,14 @@ class Builder:
         Feed is disabled if it has a value that doesn't match "rss" or
         "atom".  The templates will not generate the feed link in this
         particular case.
+
+    .. versionchanged:: 1.3.1b2
+        Feed can be disabled and will not be generated if the value of
+        ``feed_format`` is neither "rss" nor "atom" (case insensitive).
+
+    .. versionchanged:: 1.3.1b3
+        Pages and entries files are read only once, put in memory and
+        work from there without touching the files anymore.
     """
 
     def __init__(self, site_config, template_values=dict(), logger=None):
@@ -76,8 +84,8 @@ class Builder:
         self.site_config = site_config
         self.template_values = template_values
         self.site_config['dirs']['deploy'] = \
-            os.path.join(self.site_config['dirs']['deploy'],
-                         self.site_config['uri']['base'])
+            os.path.join(self.site_config.get('dirs').get('deploy'),
+                         self.site_config.get('uri').get('base'))
         self.logger = logger
 
         # Set locale for the site.
@@ -85,7 +93,7 @@ class Builder:
         try:
             self.current_locale = \
                 locale.setlocale(locale.LC_ALL,
-                                 self.site_config['wlocale']['locale'])
+                                 self.site_config.get('wlocale').get('locale'))
         except locale.Error:
             self.logger and self.logger.error(
                 "Unsupported locale setting")
@@ -97,7 +105,7 @@ class Builder:
         self.locale_dir = os.path.join(self_dir, 'data/locale')
         self.templates_dir = 'templates'
         self.builtin_templates_dir = \
-            os.path.join(self.site_config['dirs']['deploy'],
+            os.path.join(self.site_config.get('dirs').get('deploy'),
                          self.templates_dir)
 
         # dest. dirs.
@@ -113,38 +121,40 @@ class Builder:
 
         # Default values to override meta information
         self.meta_defaults = {
-            'author': self.site_config['info']['site_author'],
-            'category': self.site_config['presentation']['default_category'],
+            'author':
+                self.site_config.get('info').get('site_author'),
+            'category':
+                self.site_config.get('presentation').get('default_category'),
+            'language':
+                self.site_config.get('wlocale').get('language'),
         }
+
+        # Generate all entries metadata, once
+        self.entries_dict = self._gather_content_data().get('entries')
+        self.pages_dict = self._gather_content_data().get('pages')
 
     def __del__(self):
         """Destructor.  Restore the locale, if changed."""
         return locale.setlocale(locale.LC_ALL, self.old_locale)
 
-    def gen_entry(self, infile, date_format='%c'):
-        """Generate a HTML entry from its Markdown counterpart.
+    def gen_entry(self, filename, date_format='%c'):
+        """Generate a HTML entry from its markup language counterpart.
 
-        :param infile: Markdown file to parse
-        :type infile: str
+        :param filename: Markdown or reStructuredText file to parse
+        :type filename: str
         :param date_format: Date format for entry
         :type date_format: str
         :return: Generated HTML
         :rtype: str
         """
-        meta = self._fetch_meta(self.entries_dir, infile, True)
-        html = self._fetch_html(self.entries_dir, infile)
-
-        override = {
-            'category_uri': self._make_root_uri(meta.category(),
-                                                self.categories_dir),
-            'content': html,
-            'site_comments': self.site_config['presentation']['comments'],
-        }
+        override = {'content': self._fetch_html(self.entries_dir, filename)}
+        self.entries_dict.get(filename).update(override)
         values = self.template_values.copy()
-        values['entry'] = meta.as_dict(override, self.meta_defaults,
-                                       date_format)
+        values['entry'] = self.entries_dict.get(filename)
+        self._update_meta_date_format(values.get('entry'), date_format)
         outfile = self._make_output_file(
-            values['entry']['title'], self._entry_link_prefix(infile))
+            values.get('entry').get('title'),
+            self._entry_link_prefix(filename))
 
         return self._render_template('entry.html.j2', outfile, values)
 
@@ -154,9 +164,8 @@ class Builder:
         :param date_format: Date format for entry
         :type date_format: str
         """
-        for filename in os.listdir(self.entries_dir):
-            if valid_ext(filename):
-                self.gen_entry(filename, date_format=date_format)
+        for filename, meta in self.entries_dict.items():
+            self.gen_entry(filename, date_format=date_format)
 
     def gen_home(self, max_entries_per_page=10, date_format='%Y-%m-%d'):
         """Generate home page, and subpages.
@@ -168,34 +177,25 @@ class Builder:
         """
         entries = list()
         total_entries = 0
-        for filename in os.listdir(self.entries_dir):
-            if valid_ext(filename):
-                meta = self._fetch_meta(self.entries_dir, filename, True)
 
-                override = {
-                    'category_uri': self._make_root_uri(meta.category(),
-                                                        self.categories_dir),
-                    'odate_idx': meta.odate('%Y-%m-%d %H:%M:%S'),
-                    'uri': self._make_entry_uri(meta.title(),
-                                                filename,
-                                                absolute=True),
-                }
+        for filename, meta in self.entries_dict.items():
+            override = {'uri': self._make_entry_uri(meta.get('title'),
+                                                    filename,
+                                                    absolute=True)}
+            self.entries_dict[filename].update(override)
+            self._update_meta_date_format(meta, date_format)
+            if not meta['private']:
+                total_entries += 1
+                entries.append(self.entries_dict[filename])
 
-                # Generate entries list
-                if not meta.private():
-                    total_entries += 1
-                    entries.append(meta.as_dict(override,
-                                                self.meta_defaults,
-                                                date_format))
-
-        # Sort chronologically descent
-        entries = sorted(entries, key=lambda k: k['odate_idx'],
+        entries = sorted(entries, key=lambda k: k.get('odate_idx'),
                          reverse=True)
 
+        # Paginator
         total_pages = ceil(total_entries / max_entries_per_page)
         values = self.template_values.copy()
 
-        # FIXME: Generate 'index.html' even when there are no posts
+        # Generate 'index.html' even when there are no posts
         if total_entries == 0:
             outfile = self._make_output_file()
             self._render_template('entries.html.j2', outfile, values)
@@ -216,6 +216,7 @@ class Builder:
                 # These are the subsequent pages other than the first
                 outfile = self._make_output_file(
                     str(cur_page), self.home_cont_dir)
+
             self._render_template('entries.html.j2', outfile, values)
             values['entries'] = {}  # reset the entries dict.
 
@@ -226,38 +227,25 @@ class Builder:
         :type date_format: str
         """
         archive = dict()
-        for filename in os.listdir(self.entries_dir):
-            if valid_ext(filename):
-                meta = self._fetch_meta(self.entries_dir, filename, True)
+        for filename, meta in self.entries_dict.items():
+            override = {'uri': self._make_entry_uri(meta.get('title'),
+                                                    filename,
+                                                    absolute=False)}
+            self.entries_dict.get(filename).update(override)
+            if not meta.get('private'):
+                if meta.get('oyear_idx') in archive:
+                    update_list_in_dict(archive[meta.get('oyear_idx')],
+                                        meta.get('omonth_idx'), meta)
 
-                override = {
-                    'category_uri': self._make_root_uri(meta.category(),
-                                                        self.categories_dir),
-                    'uri': self._make_entry_uri(meta.title(),
-                                                filename),
-                    'odate_idx': meta.odate('%Y-%m-%d'),
-                }
-
-                # Generate archive
-                if not meta.private():
-                    val = meta.as_dict(override, self.meta_defaults,
-                                       date_format)
-                    oyear_idx = meta.odate('%Y')
-                    # '%m-%d' to sort chronologically
-                    omonth_idx = meta.odate('%m-%d')
-                    if oyear_idx in archive:
-                        if omonth_idx in archive[oyear_idx]:
-                            archive[oyear_idx][omonth_idx].append(val)
-                        else:
-                            archive[oyear_idx][omonth_idx] = [val]
-                        # sort entries
-                        archive[oyear_idx][omonth_idx] = \
-                            sorted(archive[oyear_idx][omonth_idx],
-                                   key=lambda k: k['odate_idx'],
-                                   reverse=False)
-                    else:
-                        archive[oyear_idx] = dict()
-                        archive[oyear_idx][omonth_idx] = [val]
+                    # sort entries
+                    archive[meta.get('oyear_idx')][meta.get('omonth_idx')] = \
+                        sorted(archive[meta.get('oyear_idx')][meta.get('omonth_idx')],
+                               key=lambda k: k.get('odate_idx'),
+                               reverse=False)
+                else:
+                    archive[meta.get('oyear_idx')] = dict()
+                    archive[meta.get('oyear_idx')][meta.get('omonth_idx')] = \
+                        [meta]
 
         values = self.template_values.copy()
         values['archive'] = archive
@@ -270,33 +258,24 @@ class Builder:
         :param date_format: Date format for entry
         :type date_format: str
         """
-        archive = dict()
-        for filename in os.listdir(self.entries_dir):
-            if valid_ext(filename):
-                meta = self._fetch_meta(self.entries_dir, filename, True)
+        categories = dict()
+        for filename, meta in self.entries_dict.items():
+            self._update_meta_date_format(meta, date_format)
+            override = {'uri': self._make_entry_uri(meta.get('title'),
+                                                    filename,
+                                                    absolute=False)}
+            self.entries_dict.get(filename).update(override)
+            if not meta.get('private'):
+                update_list_in_dict(categories, meta.get('category'), meta)
 
-                override = {
-                    'uri': self._make_entry_uri(meta.title(), filename),
-                    'odate_idx': meta.odate('%Y-%m-%d'),
-                }
-
-                # Generate archive
-                if not meta.private():
-                    val = meta.as_dict(override, self.meta_defaults,
-                                       date_format)
-                    category = meta.category()
-                    if category in archive:
-                        archive[category].append(val)
-                    else:
-                        archive[category] = [val]
-                    # sort entries
-                    archive[category] = \
-                        sorted(archive[category],
-                               key=lambda k: k['odate_idx'],
-                               reverse=True)
+                # sort entries
+                categories[meta.get('category')] = \
+                    sorted(categories[meta.get('category')],
+                           key=lambda k: k.get('odate_idx'),
+                           reverse=True)
 
         values = self.template_values.copy()
-        values['archive'] = archive
+        values['categories'] = categories
         outfile = self._make_output_file('', self.categories_dir)
         return self._render_template('catlist.html.j2', outfile, values)
 
@@ -307,33 +286,23 @@ class Builder:
         :type date_format: str
         """
         entries_by_category = dict()
-        for filename in os.listdir(self.entries_dir):
-            if valid_ext(filename):
-                meta = self._fetch_meta(self.entries_dir, filename, True)
-
-                override = {
-                    'odate_idx': meta.odate('%Y-%m-%d'),
-                    'uri': self._make_entry_uri(meta.title(), filename),
-                }
-
-                # Generate dictionary of entries by category
-                if not meta.private():
-                    val = meta.as_dict(override, self.meta_defaults,
-                                       date_format)
-                    category = meta.category()
-                    if category in entries_by_category:
-                        entries_by_category[category].append(val)
-                    else:
-                        entries_by_category[category] = [val]
+        for filename, meta in self.entries_dict.items():
+            self._update_meta_date_format(meta, date_format)
+            if not meta.get('private'):
+                self._update_meta_date_format(meta, date_format)
+                update_list_in_dict(entries_by_category,
+                                    meta.get('category'), meta)
 
         # One page for each category
         values = self.template_values.copy()
         for category in entries_by_category:
             values['category_name'] = category
-            values['entries'] = entries_by_category[category]
+            values['entries'] = entries_by_category.get(category)
+
             # sort entries
-            values['entries'] = sorted(values['entries'], key=lambda k:
-                                       k['odate_idx'], reverse=True)
+            values['entries'] = sorted(values.get('entries'),
+                                       key=lambda k: k.get('odate_idx'),
+                                       reverse=True)
             outfile = self._make_output_file(category, self.categories_dir)
             self._render_template('cat.html.j2', outfile, values)
 
@@ -344,33 +313,21 @@ class Builder:
         :type date_format: str
         """
         entries_by_tag = dict()
-        for filename in os.listdir(self.entries_dir):
-            if valid_ext(filename):
-                meta = self._fetch_meta(self.entries_dir, filename, True)
-
-                override = {
-                    'odate_idx': meta.odate('%Y-%m-%d'),
-                    'uri': self._make_entry_uri(meta.title(), filename),
-                }
-
-                # Generate dictionary of entries by tag
-                if not meta.private():
-                    val = meta.as_dict(override, self.meta_defaults,
-                                       date_format)
-                    for tag in meta.tag_list():
-                        if tag in entries_by_tag:
-                            entries_by_tag[tag].append(val)
-                        else:
-                            entries_by_tag[tag] = [val]
+        for filename, meta in self.entries_dict.items():
+            if not meta.get('private'):
+                for tag in meta.get('tag_list'):
+                    self._update_meta_date_format(meta, date_format)
+                    update_list_in_dict(entries_by_tag, tag, meta)
 
         # One page for each tag
         values = self.template_values.copy()
         for tag in entries_by_tag:
             values['tag_name'] = tag
-            values['entries'] = entries_by_tag[tag]
+            values['entries'] = entries_by_tag.get(tag)
             # sort entries
-            values['entries'] = sorted(values['entries'], key=lambda k:
-                                       k['odate_idx'], reverse=True)
+            values['entries'] = sorted(values.get('entries'),
+                                       key=lambda k: k.get('odate_idx'),
+                                       reverse=True)
             outfile = self._make_output_file(tag, self.tags_dir)
             self._render_template('tag.html.j2', outfile, values)
 
@@ -382,32 +339,28 @@ class Builder:
         will be displayed.
         """
         entries_by_tag = dict()
-        for filename in os.listdir(self.entries_dir):
-            if valid_ext(filename):
-                meta = self._fetch_meta(self.entries_dir, filename, True)
-
-                # Generate dictionary of entries by tag
-                if not meta.private():
-                    val = meta.as_dict({}, self.meta_defaults)
-                    for tag in meta.tag_list():
-                        if tag in entries_by_tag:
-                            entries_by_tag[tag].append(val)
-                        else:
-                            entries_by_tag[tag] = [val]
+        for filename, meta in self.entries_dict.items():
+            override = {'uri': self._make_entry_uri(meta.get('title'),
+                                                    filename,
+                                                    absolute=False)}
+            self.entries_dict.get(filename).update(override)
+            if not meta.get('private'):
+                for tag in meta.get('tag_list'):
+                    update_list_in_dict(entries_by_tag, tag, meta)
 
         # Multipliers seq. for tag size in function of times repeated
         tagcloud_seq = [0, 14, 21, 27, 32, 38, 42, 45, 47, 48, 50, 52]
 
         # One page for each tag
         values = self.template_values.copy()
-        values['tags'] = []
+        values['tags'] = list()
         for tag in entries_by_tag:
             if tag:
-                tagname, tagfreq = tag, len(entries_by_tag[tag])
-                mult = 100 + int(tagcloud_seq.last
+                tagfreq = len(entries_by_tag.get(tag))
+                mult = 100 + int(tagcloud_seq[-1]
                                  if tagfreq > len(tagcloud_seq)
                                  else tagcloud_seq[tagfreq - 1])
-                values['tags'].append({tagname: mult})
+                values.get('tags').append({tag: mult})
 
         outfile = self._make_output_file('', self.tags_dir)
         self._render_template('tagcloud.html.j2', outfile, values)
@@ -423,51 +376,40 @@ class Builder:
             Since this add new data to the base template, it **must be
             invoked first**, before generating any other content.
         """
-        for filename in os.listdir(self.pages_dir):
-            if valid_ext(filename):
-                meta = self._fetch_meta(self.pages_dir, filename, False)
+        for filename, meta in self.pages_dict.items():
+            if meta.get('navigation') and not meta.get('private'):
+                values = self.template_values.copy()
+                values['page'] = self.pages_dict.get(filename)
 
-                override = {'uri': self._make_root_uri(meta.title())}
+                # Update base template values to get pages links in
+                # nav. bar, only for those pages that have the meta tag
+                # ``navigation`` set to true
+                self.template_values['blog']['page_links'].append(
+                    [values.get('page').get('title'),
+                     values.get('page').get('uri')])
 
-                if meta.navigation():
-                    values = self.template_values.copy()
-                    values['page'] = meta.as_dict(override,
-                                                  self.meta_defaults)
+                self.logger and self.logger.info(
+                    'Added page to navigation: "{}"'.format(filename))
 
-                    # Update base template values to get pages links in
-                    # nav. bar only if the page is set to be in the
-                    # navigation bar
-                    self.template_values['blog']['page_links'].append(
-                        [values['page']['title'],
-                         values['page']['uri']])
+    def gen_page(self, filename):
+        """Generate a HTML page from its markup language counterpart.
 
-                    self.logger and self.logger.info(
-                        'Added page to navigation: "{}"'.format(filename))
-
-    def gen_page(self, infile):
-        """Generate a HTML page from its Markdown counterpart.
-
-        :param infile: Markdown file to parse
-        :type infile: str
+        :param filename: Markdown or reStructuredText file to parse
+        :type filename: str
         :return: Generated HTML
         :rtype: str
         """
-        meta = self._fetch_meta(self.pages_dir, infile, False)
-        html = self._fetch_html(self.pages_dir, infile)
-
-        override = {'content': html}
-
+        override = {'content': self._fetch_html(self.pages_dir, filename)}
+        self.pages_dict.get(filename).update(override)
         values = self.template_values.copy()
-        values['page'] = meta.as_dict(override, self.meta_defaults)
-
-        outfile = self._make_output_file(meta.title())
+        values['page'] = self.pages_dict.get(filename)
+        outfile = self._make_output_file(values.get('page').get('title'))
         return self._render_template('page.html.j2', outfile, values)
 
     def gen_pages(self):
         """Generate all pages."""
-        for filename in os.listdir(self.pages_dir):
-            if valid_ext(filename):
-                self.gen_page(filename)
+        for filename, meta in self.pages_dict.items():
+            self.gen_page(filename)
 
     def gen_feed(self, feed_format="atom", outfile='feed.xml'):
         """Generate blog feed.
@@ -484,90 +426,94 @@ class Builder:
 
         feed = FeedGenerator()
         # feed.logo()
-        feed.id(self.site_config['info']['site_name']
-                if self.site_config['info']['site_name']
-                else self.site_config['uri']['canonical'])
-        feed.title(self.site_config['info']['site_name']
-                   if self.site_config['info']['site_name']
-                   else self.site_config['uri']['canonical'])
-        feed.subtitle(self.site_config['info']['site_description']
-                      if self.site_config['info']['site_description']
+        feed.id(self.site_config.get('info').get('site_name')
+                if self.site_config.get('info').get('site_name')
+                else self.site_config.get('uri').get('canonical'))
+        feed.title(self.site_config.get('info').get('site_name')
+                   if self.site_config.get('info').get('site_name')
+                   else self.site_config.get('uri').get('canonical'))
+        feed.subtitle(self.site_config.get('info').get('site_description')
+                      if self.site_config.get('info').get('site_description')
                       else 'Feed')
-        feed.author({'name': self.site_config['info']['site_author'],
-                     'email': self.site_config['info']['site_author_email']})
-        feed.description(self.site_config['info']['site_description'])
-        feed.link(href=os.path.join(self.site_config['uri']['canonical'],
-                                    self.site_config['uri']['base']),
-                  rel='alternate')
-        feed.link(href=os.path.join(self.site_config['uri']['canonical'],
-                                    self.site_config['uri']['base'],
-                                    outfile), rel='self')
-        feed.language(self.site_config['wlocale']['language'])
-        feed.copyright(self.site_config['info']['copyright'])
+        feed.author({
+            'name':
+                self.site_config.get('info').get('site_author'),
+            'email':
+                self.site_config.get('info').get('site_author_email')
+        })
+        feed.description(self.site_config.get('info').get('site_description'))
+        feed.link(href=os.path.join(
+            self.site_config.get('uri').get('canonical'),
+            self.site_config.get('uri').get('base')),
+            rel='alternate')
+        feed.link(href=os.path.join(
+            self.site_config.get('uri').get('canonical'),
+            self.site_config.get('uri').get('base'),
+            outfile), rel='self')
+        feed.language(self.site_config.get('wlocale').get('language'))
+        feed.copyright(self.site_config.get('info').get('copyright'))
 
         entries = list()
-        for filename in os.listdir(self.entries_dir):
-            if valid_ext(filename):
-                meta = self._fetch_meta(self.entries_dir, filename, True)
-                html = self._fetch_html(self.entries_dir, filename)
-
-                uri = self._make_entry_uri(meta.title(), filename)
+        for filename, meta in self.entries_dict.items():
+            if not meta.get('private'):
+                uri = self._make_entry_uri(meta.get('title'), filename)
                 override = {
-                    'content_html': html,
-                    'full_uri': os.path.join(self.site_config['uri']['canonical'],
-                                             self.site_config['uri']['base'], uri),
-                    'mdate_idx': meta.mdate('%Y-%m-%d %H:%M:%S'),
-                    'odate_idx': meta.odate('%Y-%m-%d %H:%M:%S'),
-                    'timezone': 'UTC' if not meta.odate('%Z')
-                    else meta.odate('%Z'),
-                    'uri': uri,
+                    'content':
+                        self._fetch_html(self.entries_dir, filename),
+                    'full_uri':
+                        os.path.join(
+                            self.site_config.get('uri').get('canonical'),
+                            self.site_config.get('uri').get('base'),
+                            uri),
                 }
-
-                if not meta.private():
-                    val = meta.as_dict(override, self.meta_defaults)
-                    entries.append(val)
+                self.entries_dict.get(filename).update(override)
+                entries.append(meta)
 
         # Sort chronologically descent
-        entries = sorted(entries, key=lambda k: k['odate_idx'],
+        entries = sorted(entries, key=lambda k: k.get('odate_idx'),
                          reverse=True)
         for entry in entries:
             fnew = feed.add_entry()
-            fnew.id(slugify(strip_html_tags(entry['title'])))
-            fnew.title(entry['title'])
-            fnew.description(entry['content_html'])
-            if entry['mdate_html']:
-                fnew.updated(entry['mdate_html'])
-                fnew.published(entry['mdate_html'])
+            fnew.id(slugify(strip_html_tags(entry.get('title'))))
+            fnew.title(entry.get('title'))
+            fnew.description(entry.get('content'))
+            if entry.get('mdate_html'):
+                fnew.updated(entry.get('mdate_html'))
+                fnew.published(entry.get('mdate_html'))
             else:
-                fnew.updated(entry['odate_html'])
-            fnew.pubDate(entry['odate_html'])
-            # , 'email':entry['email']})
-            fnew.author({'name': entry['author']})
-            fnew.link(href=entry['full_uri'], rel='alternate')
+                fnew.updated(entry.get('odate_html'))
+            fnew.pubDate(entry.get('odate_html'))
+            # , 'email':entry.get('email')})
+            fnew.author({'name': entry.get('author')})
+            fnew.link(href=entry.get('full_uri'), rel='alternate')
 
         if feed_format.lower() == "rss":
-            feed.rss_file(os.path.join(self.site_config['dirs']['deploy'],
-                                       outfile))
+            feed.rss_file(
+                os.path.join(
+                    self.site_config.get('dirs').get('deploy'), outfile))
         elif feed_format.lower() == "atom":
-            feed.atom_file(os.path.join(self.site_config['dirs']['deploy'],
-                                        outfile))
+            feed.atom_file(
+                os.path.join(
+                    self.site_config.get('dirs').get('deploy'), outfile))
 
     def gen_static(self):
         """Generate (copies) static directory."""
         src = self.static_dir
-        dst = os.path.join(self.site_config['dirs']['deploy'],
-                           self.static_dir)
+        dst = os.path.join(
+            self.site_config.get('dirs').get('deploy'),
+            self.static_dir)
         if os.path.exists(src):
             distutils.dir_util.copy_tree(src, dst, update=True,
                                          verbose=True)
 
     def gen_extra_dirs(self):
         """Generate extra directories if they exist."""
-        if self.site_config['dirs']['extra']:
-            for extra_dir in self.site_config['dirs']['extra']:
+        if self.site_config.get('dirs').get('extra'):
+            for extra_dir in self.site_config.get('dirs').get('extra'):
                 src = extra_dir
-                dst = os.path.join(self.site_config['dirs']['deploy'],
-                                   extra_dir)
+                dst = \
+                    os.path.join(self.site_config.get('dirs').get('deploy'),
+                                 extra_dir)
                 if os.path.exists(src):
                     distutils.dir_util.copy_tree(src, dst, update=True,
                                                  verbose=True)
@@ -583,20 +529,109 @@ class Builder:
         self.logger and self.logger.info('Building static website...')
 
         self.gen_nav_page_links()
-        self.gen_entries(self.site_config['date_format']['entry'])
+        self.gen_entries(self.site_config.get('date_format').get('entry'))
         self.gen_pages()
-        self.gen_archive(self.site_config['date_format']['list'])
-        self.gen_categories(self.site_config['date_format']['list'])
-        self.gen_category_list(self.site_config['date_format']['list'])
-        self.gen_tags(self.site_config['date_format']['list'])
+        self.gen_archive(self.site_config.get('date_format').get('list'))
+        self.gen_categories(self.site_config.get('date_format').get('list'))
+        self.gen_category_list(self.site_config.get('date_format').get('list'))
+        self.gen_tags(self.site_config.get('date_format').get('list'))
         self.gen_tag_cloud()
-        self.gen_home(self.site_config['presentation']['max_entries'],
-                      self.site_config['date_format']['home'])
-        self.gen_feed(self.site_config['presentation']['feed_format'])
+        self.gen_home(self.site_config.get('presentation').get('max_entries'),
+                      self.site_config.get('date_format').get('home'))
+        self.gen_feed(self.site_config.get('presentation').get('feed_format'))
         self.gen_static()
         self.gen_extra_dirs()
 
         self.logger and self.logger.info('Done!')
+
+    def _gather_content_data(self):
+        """Gather all metadata from all parseable files.
+
+        A parseable files is one with Markdown or reStructuredText valid
+        extension, as described in :func:`fileman.has_extension_md_rst`.
+        This way, the files will be transversed only once, and stored.
+
+        This stores only entries information, not pages metadata, since
+        pages only need to be renderer "as is".  Entries, on the other
+        side, are also required to extract information in order to make
+        category pages, tags, tag clouds, etc.  So, it's needed to have
+        all that data in the same place.
+
+        The structure of the return value is as follows::
+
+            data = {
+                ['entries']: {
+                    'entry1.filename': Meta(entry1).as_dict(),
+                    'entry2.filename': Meta(entry2).as_dict(),
+                }
+                ['pages']: {
+                    'page1.filename': Meta(page1).as_dict(),
+                    'page2.filename': Meta(page2).as_dict(),
+                }
+            }
+
+        :return: Dictionary of all parseabe objects and their metadata
+        :rtype: dict
+        """
+        # Gather entries
+        entries_dict = dict()
+        for filename in os.listdir(self.entries_dir):
+            if has_extension_md_rst(filename):
+                meta = self._fetch_meta(self.entries_dir, filename,
+                                        odate_required=True)
+
+                override_entry = {
+                    'category_uri':
+                        self._make_root_uri(meta.category(),
+                                            self.categories_dir),
+                    'odate_idx': meta.odate('%Y-%m-%d %H:%M:%S'),
+                    'oyear_idx': meta.odate('%Y'),
+                    'omonth_idx': meta.odate('%m-%d'),
+                    'site_comments':
+                        self.site_config.get('presentation').get('comments'),
+                    'uri':
+                        self._make_entry_uri(meta.title(),
+                                             filename,
+                                             absolute=True),
+                }
+                entries_dict[filename] = \
+                    meta.as_dict(override_entry, self.meta_defaults)
+
+        # Gather pages
+        pages_dict = dict()
+        for filename in os.listdir(self.pages_dir):
+            if has_extension_md_rst(filename):
+                meta = self._fetch_meta(self.pages_dir, filename,
+                                        odate_required=False)
+
+                override_page = {'uri': self._make_root_uri(meta.title())}
+                pages_dict[filename] = \
+                    meta.as_dict(override_page, self.meta_defaults)
+
+        return {'entries': entries_dict, 'pages': pages_dict}
+
+    def _update_meta_date_format(self, meta, date_format):
+        """Update the date format from a meta dictionary object.
+
+        Operates on a dictionary of metainformation, output of
+        :fun:`Meta.as_dict``, so the key names are known.  Uses the
+        two datetime information fields, ``odate_info`` for the original
+        date, and ``mdate_info`` for the modified date, and rewrites in
+        place the values of their string counterparts, ``odate`` and
+        ``mdate``, if they are not ``None``.
+
+        :param meta: Metainformation dictionary
+        :type meta: dict
+        :param date_format: New date format for ``odate`` and ``mdate``
+        :type date_format: str
+
+        .. seealso:: :fun:`Meta.as_dict`
+        """
+        if meta.get('odate'):
+            meta['odate'] = meta.get('odate_info').strftime(date_format)
+
+        if meta.get('mdate'):
+            meta['mdate'] = meta.get('mdate_info').strftime(date_format)
 
     def _render_template(self, template, output_data, values):
         """Render a template using Jinja2.
@@ -611,8 +646,9 @@ class Builder:
         trans = gettext.translation('default', self.locale_dir,
                                     [self.current_locale])
         env = Environment(extensions=['jinja2.ext.i18n'],
-                          loader=FileSystemLoader([self.templates_dir,
-                                                   self.builtin_templates_dir]))
+                          loader=FileSystemLoader(
+                              [self.templates_dir,
+                               self.builtin_templates_dir]))
         env.install_gettext_translations(trans)
         env.globals['slugify'] = slugify  # Add `slugify` to Jinja2
         env.globals['strip_html_tags'] = strip_html_tags
@@ -622,14 +658,15 @@ class Builder:
         # Update only those files that are different in content
         # comparing with a cache file
         with open(output_data + '~', mode="w",
-                  encoding=self.site_config['wlocale']['encoding']) \
+                  encoding=self.site_config.get('wlocale').get('encoding')) \
                 as cache_file:
             cache_file.write(html)
 
         if not os.path.exists(output_data) or \
            not filecmp.cmp(output_data + '~', output_data):
-            with open(output_data, mode="w",
-                      encoding=self.site_config['wlocale']['encoding']) \
+            with open(
+                output_data, mode="w",
+                encoding=self.site_config.get('wlocale').get('encoding')) \
                     as output_file:
                 output_file.write(html)
             self.logger and self.logger.info(
@@ -641,55 +678,55 @@ class Builder:
 
         return html
 
-    def _fetch_markup(self, directory, infile):
+    def _fetch_markup(self, directory, filename):
         """Parse an input file depending on its extension.
 
-        The ``Parser`` constructor detects the extension of ``infile``
+        The ``Parser`` constructor detects the extension of ``filename``
         and parses its body content as well as its metadata.
 
         :param directory: Path to the file to be parsed
         :type directory: str
-        :param infile: File to parse
-        :type infile: str
+        :param filename: File to parse
+        :type filename: str
         :return: Parser object
         :rtype: Parser
 
         .. sealso:: :class:`Parser`
         """
-        return Parser(os.path.join(directory, infile),
-                      encoding=self.site_config['wlocale']['encoding'],
+        return Parser(os.path.join(directory, filename),
+                      encoding=self.site_config.get('wlocale').get('encoding'),
                       logger=self.logger)
 
-    def _fetch_html(self, directory, infile):
-        """Fetch HTML content out of a Markdown input file.
+    def _fetch_html(self, directory, filename):
+        """Fetch HTML content out of a markup language input file.
 
-        :param directory: Markdown file working directory
+        :param directory: Markdown or reStructuredText file directory
         :type directory: str
-        :param infile: Markdown file to parse
-        :type infile: str
+        :param filename: Markdown or reStructuredText file to parse
+        :type filename: str
         :return: HTML content for the parsed file
         :rtype: str
         """
-        return self._fetch_markup(directory, infile).html()
+        return self._fetch_markup(directory, filename).html()
 
-    def _fetch_meta(self, directory, infile, odate_required=False):
-        """Fetch metadata out of a Markdown input file.
+    def _fetch_meta(self, directory, filename, odate_required=False):
+        """Fetch metadata out of a markup language input file.
 
         If ``odate_required`` is ``False``, the metadata is valid as
         long as it has a title; if it's ``True``, the metadata require
         to be valid, also an "origin date", or ``odate`` field.
 
-        :param directory: Markdown file working directory
+        :param directory: Markdown or reStructuredText file directory
         :type directory: str
-        :param infile: Markdown file to parse
-        :type infile: str
+        :param filename: Markdown or reStructuredText file to parse
+        :type filename: str
         :param odate_required: Field ``odate`` is required
         :type odate_required: bool
-        :return: Markdown metadata
+        :return: Parsed file metadata
         :rtype: Meta
         """
-        return Meta(self._fetch_markup(directory, infile).metadata(),
-                    infile, odate_required, logger=self.logger)
+        return Meta(self._fetch_markup(directory, filename).metadata(),
+                    filename, odate_required, logger=self.logger)
 
     def _entry_link_prefix(self, entry):
         """Compute entry final path.
@@ -707,10 +744,11 @@ class Builder:
         .. versionchanged:: 1.3.0a
             Set as a member method.
         """
-        meta = self._fetch_meta(self.entries_dir, entry, True)
+        meta = self._fetch_meta(self.entries_dir, entry,
+                                odate_required=True)
 
         category = meta.category(
-            self.site_config['presentation']['default_category'])
+            self.site_config.get('presentation').get('default_category'))
         odate = meta.odate('%Y-%m-%d')
         odate_arr = odate.split('-')
         path = os.path.join(str(self.entries_dir),
@@ -740,7 +778,7 @@ class Builder:
         """
         if absolute:
             path = os.path.join('/',
-                                self.site_config['uri']['base'],
+                                self.site_config.get('uri').get('base'),
                                 self._entry_link_prefix(infix))
 
         else:
@@ -768,7 +806,7 @@ class Builder:
         """
         return link_to(name,
                        os.path.join('/',
-                                    self.site_config['uri']['base'],
+                                    self.site_config.get('uri').get('base'),
                                     infix),
                        makedirs=False, justdir=True, index=index)
 
@@ -787,7 +825,6 @@ class Builder:
         :rtype: str
         """
         return link_to(name,
-                       os.path.join(self.site_config['dirs']['deploy'],
+                       os.path.join(self.site_config.get('dirs').get('deploy'),
                                     infix),
                        makedirs=True, justdir=False, index=index)
-