+++ pynfact/parsers/__init__.py
@@ -41,4 +41,3 @@ __status__ = "Production"
 
 from pynfact.parsers.parserrst import ParserRst
 from pynfact.parsers.parsermd import ParserMd
-